Fix key events. More libtcod fns. Better WorldGame, faster worldgen.
Using libtcod noise instead of my homegrown code is much faster.
demo_test.py actually generates a view of the WorldGenerator heightmap,
although drawing is slow tcod.Console.__setitem__ and lots of looping
in draw_elevation are to blame.

The solution is going to be to rearchitect tcod.Console to work off
numpy arrays as libtcod.console_fill_X calls.