Integrating GameClock, new interactive consoles
Use the magic of GameClock in RootConsole main loop. New class
`InteractiveRootConsole` that plays well with ipykernel.

Intent is to slowly undo all the complexity in the GUI code (eventloop
in particular), and in the process make it easier to use interactively.

A fix for libtcod interface that finally gets console_put_char_ex
working.